Remove some hidden text
The home page has a heading that reads "My custom search widget".  The heading is white text on a white background so it's hidden from users.

This change keeps the `<h1>` tag but changes the text to a simple `$nbsp;`.  It might be safe to remove the line completely but I do not have a way to run the site locally at the moment.

Signed-off-by: Nathen Harvey <nathenharvey@google.com>